Beta (#567)
* Mastery filter & mission cost
Also bug fixes with translation circuit.  Now properly processes {{:}} as 'global.colon'.

* Add Intel to farm table for active event

* code cleanup.

* Bug fixes and optimizations.

* change cadence to a map

* Fix equipment level and item dupe bugs

* Bug fixes related to non-player event planner
Also placed a player data requirement around factions (as it apparently does need it).

Also added a more complete objective_event model

* oepicker stub

* Create new Objective Event page

* Fixing farm table mission formatting
Remove artifact chance from faction heading
Add true chance for each item to item description.
Add arrow to emphasize parent/child relationship.
Code cleanup.

* Adapt FarmTable to SearchableTable

* layout cleanup.

* Fix how jumping to collections works.

* code cleanup.

* More substantial dev work on objective events

* Finish Objective Events viewer main code

* Add "Alternative Layout" for base layout.
Collapses skills to a single column to make far columns easier to see

* remove broken link from fleet and add ColorName

* add "online since" to fleet

* Globalize the crew excluder

* Fancy excluder quick note
Intended use case:

I have 6 crew I want to use with Q missions, everyone else is inconsequential.  But I don't want to exclude all quipped crew after I've determined these six.  I just want to mark them, and then restore them later.

* Make noting exclusions additive.

* Make excluder dbid-specific

* Hover display bug

* Globalize the HOF
Also add "Last Updated" and fixed various formatting

* Add normalized avg dur cnt

* Add norm_max calc flavor

* Add event types to stat trends.

* add series + event type option

* Fix Crew view to consider subsequent milestones

* Add filtered milestone reward hover

* Visual improvement to hover

* rename BuffBase -> ItemArchetypeBase

* Refactor to consolidate EquipmentCommon/EquipmentItem

* Merge Icon and AtlasIcon

* fix filter to use extendedCollections in crew mode

* make sure the selection is sane

* bug

* Move ObjectiveEvent model back to player.ts
For compatibility with other tools

* Refactor player glance in a sane way
Prevents double rendering, cleanly renders with memoized data, etc.

* code cleanup.

* Add buff state for best quipment, natural/full slots

* fr/es verbiage

* Change DataScore rank to in-rarity rank
This aligns with CAB and also the bot.

* Fix multiworker
Not returning some results.

* Visual fixes of objective rewards
Also minor bug with item over and quantum

* FTM Hall of Fame

* FTMs and OEs
Add OE achievement tracking when player data is present.

Add group-by-player functionality to the FTM wall

* set a default sort for FTMs

* fix search bug

* Update ftm page to use live data.

* Add DataScore ranks to crew page and cool stats

* Add url search param support to ftm

* Add mini-helper framework to OE

* Add rarity filter

* tweak sorting.

* Fix filtering logic and no results bugs.

* add filter by skills

* Add mortal helpers

* add level helper.

* Add logic preference.

* add possible bonus to quipment data.

* Add datascore to item crew,
add quipment score to quipment item crew

* add copy to clipboard functionality.

* v3.1 Release Notes

* Stabilize calculation kick-off on initial page launch

* Bump DataScore stats to top 50.
More release notes

* Refactor of profile pages.

* Add roster summary
Also add ability to refresh page, and more efficient loading of profile data (usually just once, now)

* Rename Stat Trends

* Add ability for children to share running worker context

* refactor to provide profile page with farm sources

* (carefully) add shuttle countdown to event blurb

* bug

* More graceful timer.

* Fix formatting of negative numbers

* make shuttle update less frenetic

* use tomato

* Add progress to next column on quipment views

* client_api=27 -> QBits in stored_immortals!

* bug fix for stored immortal q-bits; announcement

* Bug fix, stored immortals qbits.

* fix preview crew showing as owned.

* Major enhancements to PVP opponents
Add ability to import current division opponents to quickly select a configuration to use.
Translations and tweaks for Fleet data and JSON import data.
Now can select from full roster for opponents as many times as you like.

* Disable clear all for selected pvp

* Fix bug with staffing changes not reflected

* Prevent doubling up of ship passives.

* add to new features.

* disable CAP achiever for preview crew.

* attempting to fix voyage aggregates copy

* v3.1 General Release

---------

+++ src/components/voyagehistory/voyagestable.tsx
@@ -3,8 +3,10 @@ import {
 	Dropdown,
 	DropdownItemProps,
 	Form,
+	Icon,
 	Message,
 	Pagination,
+	Popup,
 	Table
 } from 'semantic-ui-react';
 
@@ -15,7 +17,7 @@ import { formatTime } from '../../utils/voyageutils';
 import CONFIG from '../CONFIG';
 
 import { HistoryContext } from './context';
-import { deleteTrackedData, removeVoyageFromHistory, SyncState } from './utils';
+import { deleteTrackedData, postUnsynchronizedVoyages, removeVoyageFromHistory, SyncState } from './utils';
 import { VoyageModal } from './voyagemodal';
 
 interface ITableState {
@@ -88,7 +90,8 @@ export const VoyagesTable = () => {
 		{ /* Ship Trait */ column: '_shipTrait', title: t('voyage.voyage_history.fields.ship_trait') },
 		{ /* Antimatter */ column: 'max_hp', title: t('voyage.voyage_history.fields.antimatter'), firstSort: 'descending' },
 		{ /* Initial Estimate */ column: 'estimate.median', title: t('voyage.voyage_history.fields.initial_estimate'), firstSort: 'descending' },
-		{ /* Last Estimate */ column: 'checkpoint.estimate.median', title: t('voyage.voyage_history.fields.last_estimate'), firstSort: 'descending' }
+		{ /* Last Estimate */ column: 'checkpoint.estimate.median', title: t('voyage.voyage_history.fields.last_estimate'), firstSort: 'descending' },
+		{ /* Remotely Tracked */ column: 'remote', title: t('voyage.history.remote_tracking') }
 	];
 
 	// Filter
@@ -175,32 +178,51 @@ export const VoyagesTable = () => {
 		const dtCreated: Date = new Date(row.created_at);
 		const isRunning: boolean = row.voyage_id > 0 && !!ephemeral?.voyage.find(v => v.id === row.voyage_id);
 		return (
-			<Table.Row key={row.tracker_id} onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
-				<Table.Cell>
-					{dtCreated.toLocaleDateString()}
-					{isRunning && <><br/>{t('voyage.running_voyage')}</>}
+			<Table.Row key={row.tracker_id}>
+				<Table.Cell style={{
+					display: 'flex',
+					padding: '1em',
+					flexDirection: 'row',
+					alignItems: 'center',
+					justifyContent: 'flex-start',
+					gap: '1em'
+				}}>
+					<Icon name='trash' onClick={() => removeTrackedVoyage(row.tracker_id)} style={{cursor: 'pointer'}} />
+					<div onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
+						{dtCreated.toLocaleDateString()}
+						{isRunning && <><br/>{t('voyage.running_voyage')}</>}
+					</div>
 				</Table.Cell>
-				<Table.Cell textAlign='center'>
+				<Table.Cell textAlign='center' onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
 					{CONFIG.SKILLS[row.skills.primary_skill]}
 				</Table.Cell>
-				<Table.Cell textAlign='center'>
+				<Table.Cell textAlign='center' onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
 					{CONFIG.SKILLS[row.skills.secondary_skill]}
 				</Table.Cell>
-				<Table.Cell textAlign='center'>
+				<Table.Cell textAlign='center' onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
 					{SHIP_TRAIT_NAMES[row.ship_trait] ?? row.ship_trait}
 				</Table.Cell>
-				<Table.Cell textAlign='center'>
+				<Table.Cell textAlign='center' onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
 					{row.max_hp}
 				</Table.Cell>
-				<Table.Cell textAlign='center'>
+				<Table.Cell textAlign='center' onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
 					<b>{formatTime(row.estimate.median, t)}</b>
 					<br />({`${formatTime(row.estimate.minimum, t)} - ${formatTime(row.estimate.moonshot, t)}`})
 				</Table.Cell>
-				<Table.Cell textAlign='center'>
+				<Table.Cell textAlign='center' onClick={() => setActiveVoyage(row)} style={{ cursor: 'pointer' }}>
 					{renderLastEstimate(row.checkpoint)}
+				</Table.Cell>
+				<Table.Cell textAlign='center'>
+					{row.remote ? t('global.yes') : t('global.no')}
+					{!row.remote && <Icon onClick={() => synchronizeVoyage(row)} style={{margin: '0.5em', cursor: 'pointer'}} name='upload' />}
 				</Table.Cell>
 			</Table.Row>
 		);
+	}
+
+	async function synchronizeVoyage(row: ITrackedVoyage) {
+		await postUnsynchronizedVoyages(dbid, history, row.tracker_id);
+		setHistory({...history});
 	}
 
 	function renderLastEstimate(checkpoint: ITrackedCheckpoint): JSX.Element {
@@ -223,6 +245,10 @@ export const VoyagesTable = () => {
 	}
 
 	function removeTrackedVoyage(trackerId: number): void {
+		if (typeof window !== 'undefined') {
+			let result = window.confirm(t('voyage.history.warn_delete'));
+			if (!result) return;
+		}
 		if (syncState === SyncState.RemoteReady) {
 			deleteTrackedData(dbid, trackerId).then((success: boolean) => {
 				if (success) {
